take views out of settings
